apyCal func added
