+++ apps/wallet/src/ui/app/staking/ApyCalc.ts
@@ -1,0 +1,18 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { ActiveValidator } from './ValidatorDataTypes';
+
+export function apyCalc(validators: ActiveValidator, epoch: number) {
+    const { sui_balance, starting_epoch, delegation_token_supply } =
+        validators.fields.delegation_staking_pool.fields;
+
+    const num_epochs_participated = +epoch - +starting_epoch;
+    const apy = Math.pow(
+        1 +
+            (+sui_balance - +delegation_token_supply.fields.value) /
+                +delegation_token_supply.fields.value,
+        365 / num_epochs_participated - 1
+    );
+    return apy ? apy : 0;
+}
