+++ apps/wallet/src/ui/app/staking/ValidatorDataTypes.ts
@@ -1,10 +1,69 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-//TODO pull from the SDK
+import { type SuiAddress } from '@mysten/sui.js';
 
 export const VALIDATORS_OBJECT_ID = '0x5';
 export const VALDIATOR_NAME = /^[A-Z-_.\s0-9]+$/i;
+
+// NOTE Temporary  until SUI SDK is updated
+// TODO: add to SUI SDK once Validator types is finalized
+// Get validators by account address
+/**
+ *
+ * @see {@link https://github.com/MystenLabs/sui/blob/b904dede65c91c112434d49180e2d277e76ccee6/crates/sui-types/src/sui_system_state.rs#L42}
+ *
+ */
+
+const STAKE_DELEGATOR_STALE_TIME = 5 * 1000;
+
+export type ValidatorMetaData = {
+    sui_address: SuiAddress;
+    pubkey_bytes: number[];
+    network_pubkey_bytes: number[];
+    worker_pubkey_bytes: number[];
+    proof_of_possession_bytes: number[];
+    name: number[];
+    net_address: number[];
+    consensus_address: number[];
+    worker_address: number[];
+    next_epoch_stake: number;
+    next_epoch_delegation: number;
+    next_epoch_gas_price: number;
+    next_epoch_commission_rate: number;
+};
+
+// Staking
+type Id = {
+    id: string;
+};
+
+type Balance = {
+    value: bigint;
+};
+
+type StakedSui = {
+    id: Id;
+    validator_address: SuiAddress;
+    pool_starting_epoch: bigint;
+    delegation_request_epoch: bigint;
+    principal: Balance;
+    sui_token_lock: bigint | null;
+};
+
+type ActiveDelegationStatus = {
+    Active: {
+        id: Id;
+        staked_sui_id: SuiAddress;
+        principal_sui_amount: bigint;
+        pool_tokens: Balance;
+    };
+};
+
+export type DelegatedStake = {
+    staked_sui: StakedSui;
+    delegation_status: 'Pending' | ActiveDelegationStatus;
+};
 
 export interface Validators {
     dataType: string;
